Adding new properites to MultiplayerSession to enable update MPSD Thunderhead allocation method
Changes to support a new method of MPSD Thunderhead allocation (the same
method that PartyChat is now using). To support using the new allocation
method, I needed to make some changes to how we create and interact with
the game session doc.

We use the XSAPI MultiplayerSession^ on our game clients to create and
interact with game sessions. However, we now needed to be able to read
and write a few new things to the session doc that weren’t supported out
of the box.

1. These values must be set in the session doc at creation:

/constants/system/cloudComputePackage/titleId
: “your cloud title ID as a string”
/constants/system/cloudComputePackage/gsiSet
: “UUID of your cloud compute package”

I added an additional optional parameter to the MultiplayerSession^
constructor called systemConstants, which allows us to pass in json that
gets put into the “/constants/system/” section when creating a new game
session. (I followed the pattern of the existing optional parameter on
the MultiplayerSession^ constructor customConstants, which is json that
is written to “/constants/custom/”). Note that we also need to set these
values on our matchmaking sessions, but the existing
MultiplayerSession::SetMatchmakingTargetSessionConstantsJson() allowed
us to write them into session without any need for XSAPI changes.

2. Then when we want to request an allocation from the client, we need
to write a flag into the session doc

/properties/system/allocateCloudCompute
: Boolean

I added a property to MultiplayerSession^ called AllocateCloudCompute,
following the pattern of the existing Closed property, so that the
client could easily set this to true to request an allocation, and later
detect if MPSD set it to false again (which is how it indicates a failed
allocation).