Merged PR 660748: Remove check blocking requests NSAL doesn't require token for
Context from email discussion with Jason:
The second issue I fixed in the commit is about this line was to fix the issue in the attached “token” thread.
But this 2nd issue is no longer an issue due to the “updated CLL is available” thread, and we’ve already taken this updated CLL.
So in theory you should be able to change this line and remove this logic without issue.

What we were seeing happening here is that we would get down to auth manager and NSAL would say that a token wasn't required, so we would pass up an empty token. When we got to the user_context level, we had an explicit check that would see if we did not receive an xbox error but did receive an empty token and throw a new error.

The original reason for this workaround has since been resolved, and this logic was blocking all requests that NSAL believed to not require a token from being fired, so now we're removing this check.

I've since verified that several of the previously blocked calls are now flowing.

Related work items: #12672080