Fix for multiplayer_manager::join_loby() (#389)
We want the invited user to join first since it is possible that users in the local graph might not meet the criteria to join the invited session until the invited user joins.

Imagine that the inviting session has the session privacy set to joinable_by_friends and only the invited user is a friend of the inviting user. If any additional users attempt to join the inviting session before the invited user than the join fails for the whole list of users.